updated 2d partition

+++ 2D Partition/include/prefixSum.cuh
@@ -6,7 +6,7 @@ class PrefixSum
     public:
         void sum_scan_blelloch(int* const d_out,
                                 const int* const d_in,
-                                const size_t numElems);  // Function used by GPU
+                                const int numElems);  // Function used by GPU
 };
 
 #endif
